Remove BC keyword parsing logic
Script doesn't parse keywords anymore, but forwards BC's output directly to user's TTY.
In that way any syntax errors are immediately displayed to the user, and any further change in BC doesn't need extra work on this script.
Also, the output of infinite loops is visible by the user now.